requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
numpy==1.18.4
pandas==1.0.4
python-dateutil==2.8.1
python-dotenv==0.13.0
python-gitlab==2.2.0
pytz==2020.1
requests==2.23.0
six==1.15.0
urllib3==1.25.9
